==> Scripts/01b_filteredICs_look.Rmd <==
---
title: "01b_filteredIC_look"
author: "Christian Coffman"
date: '2022-05-04'
output: html_document
---

```{r setup, include=FALSE}
library(tidyverse)
library(fslr)
```

```{r cars}
ICs <- read.csv("../thresh50/dFNC_labels.csv")
ic_files <- data.frame(files = list.files("../SMs")) %>%
  mutate(num = parse_number(files)) %>%
  filter(num %in% ICs$IC)

```


```{r}
atlas_file = '../SMs/Yeo_MNI.nii.gz'
reffile = '/home/christian/Research/Brain_stuff/Atlases/MNI152.nii.gz'

for (i in ic_files$files) {
  filepath <- paste0("../SMs/", i)
  ortho2(reffile, filepath)
}

flirt(infile = atlas_file, reffile = reffile, dof = 6, reorient = T,
                outfile = '../SMs/Yeo_MNI.nii.gz')

```
