==> Scripts/01c_inspect_SMs.Rmd <==
---
title: "01c_inspect_SMs"
author: "Christian Coffman"
date: '2022-05-05'
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
library(tidyverse)
library(fslr)
```


```{r}
home <-  '/home/christian/Research/Brain_stuff/BRIDGES_GIG_ICA/GIG-6/'
IC_SM <- paste0(home, 'GIG/gica_cmd_mean_component_ica_s_all_.nii')

# split the aggregate map into IC's (which are "time" dimension")
IC_SMlist = fslsplit(IC_SM, direction = "t")
ICs <- read_csv(paste0(home, "selected_ICs.csv"))

ref <- readNIfTI('/home/christian/Research/Brain_stuff/Atlases/MNI152.nii.gz')
```

```{r}
for (i in 1:nrow(ICs)) {
  ICnum <- ICs$IC[i]
  ICname <- ICs$ICname[i]
  SM2 <- ifelse(IC_SMlist[[ICnum]] > quantile (IC_SMlist[[ICnum]], 0.995),
                IC_SMlist[[ICnum]], NA)

  # Find max voxel
  m <- which(SM2 == max(SM2, na.rm = T), arr.ind = T)
  ortho2(ref, SM2, crosshairs = F, xyz=  m, text = paste0("IC", ICnum, ": ", ICname))
}
```
