==> Scripts/06-fig_tabs.Rmd <==
---
title: "06-Pub_figs_tabs"
author: "Christian Coffman"
date: "11/10/2021"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(mice)
library(mitml)
library(tidyverse)
library(kableExtra)
library(lme4)
```

# Switching regression results
```{r}
imp.sw <-readRDS("/home/christian/Research/Brain_stuff/BRIDGES_GIG_ICA/GIG-ICA-5-updated-GIFT/thresh50/Book_regressions/imputed_switching_data.rds")


d1 <- d1 %>% 
  inner_join(znew, by = "grid") %>%
  filter(is.na(No_Scan), (Rest_motion_exclusion == 0 | is.na(Rest_motion_exclusion)), RS_fieldmap_exclusion == 0 ) %>% 
  select(grid, age, BDI, BSSI, Lifetime.NSSI.Group, lifetime_nssie, Freq) 
z<- read.csv("~/Research/Brain_stuff/BRIDGES_GIG_ICA/complete bridges y1.csv", header = TRUE)

znew <- z %>% 
  # filter(is.na(No_Scan), Rest_motion_exclusion !=1) %>%
  dplyr::select(grid, No_Scan, Rest_motion_exclusion, RS_fieldmap_exclusion, dropped.quit) %>%
  filter(is.na(No_Scan)) %>%
  mutate(grid = factor(grid))

dj<- znew %>% filter(Rest_motion_exclusion !=1, RS_fieldmap_exclusion != 1)


# filter out ones not in Zeynep's
imp.sw<- filter(imp.sw, grid %in% dj$grid)
fit.sw <- with(imp.sw, lm(Freq ~ age +  BSSI))
rep <- summary(pool(fit.sw))[2:3, c(1,2,3,6)]
rep %>%
  kable("html") %>%
  kable_paper()
```


```{r}

imp.dw <-readRDS("/home/christian/Research/Brain_stuff/BRIDGES_GIG_ICA/GIG-ICA-5-updated-GIFT/thresh50/Regressions/imputed_dwell_data.rds")
imp.dw<- filter(imp.dw, grid %in% dj$grid)

fit.dw <- with(imp.dw, lmer(log(Dwell.time) ~ scale(age)*State + scale(BSSI) * State + (1|grid)))

dwell_fit <- testEstimates(as.mitml.result(fit.dw), extra.pars = T)


fixed <- dwell_fit$estimates[-1, c(1,2,5)]
# rownames(fixed)[3] <- "Age"
fixed <- as.data.frame(fixed)


fixed %>%
  mutate(ID = row.names(fixed)) %>%
  separate(ID, into = c("Effector", "State"), sep = ":") %>%
  dplyr::select(Effector, State, Estimate, Std.Error, `P(>|t|)`) %>%
  mutate(Effector= str_replace(Effector, "Lifetime.NSSI.Group.L", "Mild NSSI"),
         Effector= str_replace(Effector, "Lifetime.NSSI.Group.Q", "Moderate NSSI"),
         Effector= str_replace(Effector, "Lifetime.NSSI.Group.C", "Severe NSSI"),
         State = str_remove(State, "State")) %>% kable()
  # kable(digits = c(3,3,3), row.names = FALSE, full_width = F, font_size = 12)
rands <-  dwell_fit$extra.pars
# rownames(rands) <- c("ID", "Residual","Residual")
rands %>%
  kable()
  #kable(col.names = "$\\hat\\sigma$", full_width = F, font_size = 12) 

vcov(fit.sw$analyses[[1]])
```


# Other regeessions
```{r}
switch_all <- with(imp.sw, lm(Freq ~ age +  BSSI + BDI + log(lifetime_nssie+1)))
a <- as.data.frame(summary(pool(switch_all))) %>% mutate(term = as.character(term))
a[5,1] <- "log NSSI events"
a %>%
  dplyr::select(term, estimate,std.error, p.value) %>%
  filter(term != "(Intercept)") %>% 
  kable(digits= c(4,4,4))

age <- with(imp.sw, lm(Freq ~ log(lifetime_nssie+1)))
a <- as.data.frame(summary(pool(age))) %>% mutate(term = as.character(term))
a %>%
  dplyr::select(term, estimate,std.error, p.value) %>%
  filter(term != "(Intercept)") %>% 
  kable(digits= c(4,4,4))

```


```{r}
library(mclogit)
library(MASS)
library(nnet)

df <- imp.dw$data


df$State <- relevel(df$State, ref = "1")
age_df <- read.csv("/home/christian/Research/Brain_stuff/BRIDGES_GIG_ICA/switching_data.csv") 
df <- df %>% filter(grid %in% age_df$grid)
test <- multinom(State~ age + BSSI, data = df)

z <- summary(test)$coefficients/summary(test)$standard.errors
p <- (1 - pnorm(abs(z), 0, 1)) * 2
round(p[,2:3], 4)
```


```{r}
v <- vcov(lm(Freq ~ age + BSSI + log(lifetime_nssie + 1) + BDI, data = imp.sw$data))

as.data.frame(v[-1,-1]) %>%
  mutate(age = age/  5.254327e-07,
         BSSI = BSSI / 2.362998e-08,
         `log(lifetime_nssie + 1)` = `log(lifetime_nssie + 1)` / 2.473466e-06,
         BDI = BDI / 7.925545e-09)

imp.sw$data %>%
  dplyr::select(BSSI, BDI, lifetime_nssie, age) %>%
  mutate(lifetime_nssie = log(lifetime_nssie + 1),
         age = as.numeric(age)) %>%
  cor(use = "pairwise.complete.obs")

summary(pool(with(imp.sw, lm(Freq ~ age))))$estimate
summary(pool(with(imp.sw, lm(Freq ~ BDI))))$estimate
summary(pool(with(imp.sw, lm(Freq ~ BSSI))))$estimate
summary(pool(with(imp.sw, lm(Freq ~log(lifetime_nssie+1)))))$estimate

summary(pool(with(imp.sw, lm(Freq ~ age +  BSSI+ BDI + log(lifetime_nssie+1)))))$estimate

```
