==> Scripts/09_Depression_regression.Rmd <==
---
title: "09_Depression_regression"
author: "Christian Coffman"
date: "2/16/2022"
output: pdf_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(tidyverse)
library(MASS)
library(kableExtra)
```

Load data and grids in our study
```{r cars}
diags <- read.csv("~/Research/Brain_stuff/BRIDGES_GIG_ICA/Diagnoses/ALongitudinalStudyEx_DATA_LABELS_2022-02-22_1045.csv")
# grids2 <- data.frame(grid = proj2$grid, arm = proj2$redcap_event_name)
good_grids <- read.csv("/home/christian/Research/Brain_stuff/BRIDGES_GIG_ICA/switching_data.csv")$grid

```


Look at number of depressions

1. Count every subjects that has a depression diagnosis in any of the columns. I'll call these counts "Diagnosis"
2. Filter those counts by only the ones having the word "Checked" in the columns that have current in them. An example column is  "Status.of.primary.diagnosis...choice.Current." I will not count checks in other columns such as the following:
Status.of.primary.diagnosis...choice.Past
Status.of.primary.diagnosis...choice.Partial.remission.
Status.of.primary.diagnosis...choice.Full.remission
Status.of.primary.diagnosis...choice.Recurrent
Status.of.primary.diagnosis...choice.With.psychotic.features
Status.of.primary.diagnosis...choice.Missing.Skipped
Status.of.primary.diagnosis...choice.More.than.one.response
```{r}

selects <- diags %>%
  # only grab diagnosis at year 1
  filter(grepl("Year 1", Event.Name)) %>%
  # only grab subejcts for which we have scans 
  mutate(GRID.number. = as.numeric(GRID.number.)) %>%
  filter(GRID.number. %in% good_grids) %>%
  # looks like there are 124 so that's good, keep going
  dplyr::select(contains("diagnosis") | "GRID.number.") %>%
  # Exclude other designations for the diagnosis timeline
  dplyr::select(!contains("choice.Past")) %>%
  dplyr::select(!contains("choice.Partial.remission.")) %>%
  dplyr::select(!contains("choice.Full.remission")) %>%
  dplyr::select(!contains("choice.Recurrent")) %>%
  dplyr::select(!contains("choice.With.psychotic.features")) %>%
  dplyr::select(!contains("choice.Missing.Skipped")) %>%
  dplyr::select(!contains("choice.More.than.one.response")) %>%
  dplyr::select(!contains("Severity")) %>%
  dplyr::select(!contains("specify")) %>%
  dplyr::select(!contains("choice.More.than.one.response"))
```

```{r}
current_depress <- selects %>% 
  # pivot taller
  pivot_longer(-GRID.number., names_to = "Name", values_to = "Value")  %>% 
  # mutate s.t. the names are simpler, (i.e. there are just diagnoses and current status)
  # also add keep track of diagnosis number 
  mutate(Name = ifelse(grepl("Current", Name), "Status", "Diagnosis"), Diag = rep(rep(1:9, each=2), 124)) %>%
  pivot_wider(id_cols = c(GRID.number., Diag) ,names_from = Name, values_from = Value) %>%
  filter(Diagnosis != "") %>%
  filter(Diagnosis!= "No Diagnosis", Diagnosis!= "No diagnosis", !grepl("Other", Diagnosis), !grepl("Oppositional", Diagnosis), !grepl("Obsess", Diagnosis)) %>%
  filter(!grepl("Phobia", Diagnosis), !grepl("Anxiety", Diagnosis), !grepl("Attention", Diagnosis), !grepl("Post T", Diagnosis), !grepl("Panic", Diagnosis)) %>%
  filter(!grepl("Rule", Diagnosis), !grepl("Substance", Diagnosis)) %>%
  # binarize status and depression
  mutate(Status = Status == "Checked") %>%
  # Count number of depressions per person
  group_by(GRID.number., Status) %>%
  summarize(Depress = sum(Status)) %>%
  filter(Status == TRUE, Depress >0)

# 86 with depression
# 68 current 

write.csv(current_depress, "/home/christian/Research/Brain_stuff/BRIDGES_GIG_ICA/current_depression.csv")

```

```{r}
  dplyr::select(!contains("specify")) %>%
  dplyr::select(!contains("sev")) %>%
  pivot_longer(-grid)  %>%
  mutate(d = ((value == 1) | (value == 2))) %>%
  mutate(name = ifelse(grepl("diagnosis", name), "Diag", "Status")) %>%
  group_by(grid, name) %>%
  summarise(depr = sum(d, na.rm=TRUE)) %>% 
  mutate(depr = depr > 0) 

depress %>%
  filter(grid %in% good_grids) %>% 
  group_by(name) %>%
  summarize(number = sum(depr)) %>%
  mutate(prop = number/124) %>% 
  kable()

```


```{r}



df <-readRDS("/home/christian/Research/Brain_stuff/BRIDGES_GIG_ICA/GIG-ICA-5-updated-GIFT/thresh50/Regressions/imputed_switching_data.rds")$data



temp <- depress %>% 
  pivot_wider(names_from = name, values_from = depr) %>%
  mutate(grid = factor(grid)) %>%
  inner_join(df, by = "grid") %>%
  dplyr::select(grid, Diag, Status, age, BSSI, BDI, Freq) %>%
  drop_na()



st <- lm(Freq ~ . - grid, temp)
stepAIC(st, direction = "both", trace = 0, k = log(4))#, scope = list(upper = ~ ., lower = ~ age  ))

depress %>% 
  pivot_wider(names_from = name, values_from = depr) %>%
  mutate(grid = factor(grid)) %>%
  inner_join(df, by = "grid") %>%
  dplyr::select(Diag, Status, age, BDI, Freq) %>%
  drop_na() %>%
  lm(Freq ~ Status, data=.) %>% 
  summary()


plot(best)
  

temp %>%
  ggplot(aes(x= depr, y = Freq)) +
  geom_point()
```

A look at dates

```{r}
df <- read.csv("../../../complete bridges y1.csv")

df %>% 
  dplyr::select(grid, scan_date) %>%
  mutate(date = as.Date(scan_date, format= "%m/%d/%Y")) %>%
  ggplot(aes(x= date))+
  geom_histogram() +
  geom_vline(xintercept = as.Date("2020-06-01", formate="5Y-%m-%d"), col = "red")


df %>% 
  dplyr::select(grid, scan_date) %>%
  mutate(date = as.Date(scan_date, format= "%m/%d/%Y")) %>%
  mutate(cutoff = as.Date("2020-06-01", formate="5Y-%m-%d")) %>%
  mutate(beyond = date > cutoff) %>%
  summarize(sum(beyond, na.rm= TRUE))

```

```{r}
df <- read.csv("../../../Dates_of_KSADS_and_Rs.csv")
```

```{r}
df %>%
  mutate(date=  as.Date(scan_date, format= "%m/%d/%Y"))
```
